trabalho-thiago-m/trabalho-thiago-main/Comanda-Digital/dish-app/src/app/pages/menu/menu.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { MenuComponent } from './menu.component';
import { Router } from '@angular/router';
import { Location } from '@angular/common';

describe('MenuComponent', () => {
  let component: MenuComponent;
  let fixture: ComponentFixture<MenuComponent>;
  let routerSpy = { navigate: jasmine.createSpy('navigate') };
  let locationSpy = { back: jasmine.createSpy('back') };

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [MenuComponent],
      providers: [
        { provide: Router, useValue: routerSpy },
        { provide: Location, useValue: locationSpy }
      ]
    }).compileComponents();

    fixture = TestBed.createComponent(MenuComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should navigate back when goBack is called', () => {
    component.goBack();
    expect(locationSpy.back).toHaveBeenCalled();
  });

  it('should navigate to product details page when goToProductDetails is called', () => {
    component.goToProductDetails();
    expect(routerSpy.navigate).toHaveBeenCalledWith(['/product-details']);
  });
});
